requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.35.0
torch>=2.2.0
numpy>=1.26.0
PyPDF2==3.0.1
langchain==0.0.339
faiss-cpu>=1.7.4
openai==1.3.0
tiktoken==0.5.1
openai-whisper
sounddevice
scipy
accelerate==0.24.0